Make HTTP client configurable in Emitter
