+++ snowplow-cpp-tracker.vcxproj
@@ -107,7 +107,9 @@
     <ClCompile Include="src\client_session.cpp" />
     <ClCompile Include="src\cracked_url.cpp" />
     <ClCompile Include="src\emitter.cpp" />
-    <ClCompile Include="src\http_client.cpp" />
+    <ClCompile Include="src\http_client_apple.cpp" />
+    <ClCompile Include="src\http_client_windows.cpp" />
+    <ClCompile Include="src\http_client_test.cpp" />
     <ClCompile Include="src\http_request_result.cpp" />
     <ClCompile Include="src\payload.cpp" />
     <ClCompile Include="src\self_describing_json.cpp" />
@@ -136,7 +138,10 @@
     <ClInclude Include="src\constants.hpp" />
     <ClInclude Include="src\cracked_url.hpp" />
     <ClInclude Include="src\emitter.hpp" />
-    <ClInclude Include="src\http_client.hpp" />
+    <ClInclude Include="src\ihttp_client.hpp" />
+    <ClInclude Include="src\http_client_apple.hpp" />
+    <ClInclude Include="src\http_client_windows.hpp" />
+    <ClInclude Include="src\http_client_test.hpp" />
     <ClInclude Include="src\http_request_result.hpp" />
     <ClInclude Include="src\payload.hpp" />
     <ClInclude Include="src\self_describing_json.hpp" />
